Move Helm deployment guides
Backports #13105

See #12654

The Kubernetes Access section includes guides for using the Kubernetes
Service as well as guides for running the Auth and Proxy Service on
Kubernetes. This is misleading, since (a) you can run the Auth/Proxy
on Kubernetes without using Kubernetes Access and (b) you can use
Kubernetes Access without running the Auth/Proxy on Kubernetes.

This change focuses on our Helm deployment guides. These guides
are not related to the Kubernetes Service, but rather to deploying
the Auth Service and Proxy Service on Kubernetes. I've suggested
moving these guides to a /setup/helm-deployments section for visibility.